argv.h: usage screen generator: remove the paradigm-supporting implementation.
The idea was nice, the functionality was helpful for development, yet the
resulting layout and overall appearance were rather awful. Regardless of
the effort that was put into this sophisticated flavor of the usage
screen generator, and despite it's utility for debugging, it is now
being replaced with a simpler, two-field-only variant.